requirements.txt
absl-py==1.3.0
allure-pytest==2.13.5
allure-python-commons==2.13.5
anyio==3.6.2
Appium-Python-Client==4.1.0
apturl==0.5.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.7.2
asttokens==2.2.1
astunparse==1.6.3
attrs==24.2.0
backcall==0.2.0
bcrypt==3.2.0
beautifulsoup4==4.11.1
beniget==0.4.1
bleach==6.0.0
blinker==1.4
Brlapi==0.8.3
cachetools==5.2.0
certifi==2022.12.7
cffi==1.15.1
chardet==4.0.0
charset-normalizer==2.1.1
cliapp==1.20180812.1
click==8.0.3
cmdtest==0.32+git
colorama==0.4.4
comm==0.1.2
command-not-found==0.3
contourpy==1.0.7
cryptography==3.4.8
cupshelpers==1.0
cycler==0.11.0
dbus-python==1.2.18
debugpy==1.6.6
decorator==4.4.2
defer==1.0.6
defusedxml==0.7.1
distlib==0.3.4
distro==1.7.0
distro-info==1.1+ubuntu0.2
Django==4.1.13
django-dotenv==1.4.2
django-phonenumber-field==7.1.0
djongo==1.3.6
dnspython==2.6.1
docker==5.0.3
dockerpty==0.4.1
docopt==0.6.2
duplicity==0.8.21
exceptiongroup==1.2.2
executing==1.2.0
fasteners==0.14.1
fastjsonschema==2.16.2
filelock==3.9.0
Flask==2.2.2
flatbuffers==23.1.4
fonttools==4.38.0
fqdn==1.5.1
ftfy==6.1.1
future==0.18.2
gast==0.4.0
google-auth==2.15.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.51.1
gyp==0.1
h11==0.14.0
h5py==3.7.0
html5lib==1.1
httplib2==0.20.2
idna==3.4
ifaddr==0.1.7
imageio==2.23.0
importlib-metadata==4.6.4
iniconfig==2.0.0
ipykernel==6.20.2
ipython==8.9.0
ipython-genutils==0.2.0
ipywidgets==8.0.4
isoduration==20.11.0
itsdangerous==2.1.2
jax==0.4.10
jedi==0.18.2
jeepney==0.7.1
Jinja2==3.1.2
joblib==1.2.0
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-events==0.6.3
jupyter_client==8.0.1
jupyter_core==5.1.5
jupyter_server==2.1.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.5
keras==2.12.0
keyring==23.5.0
kiwisolver==1.4.4
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
libclang==14.0.6
lockfile==0.12.2
louis==3.20.0
lxml==4.8.0
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.3
matplotlib-inline==0.1.6
mistune==2.0.4
ml-dtypes==0.1.0
mongoengine==0.29.0
monotonic==1.6
more-itertools==8.10.0
MouseInfo==0.1.3
nbclassic==0.5.1
nbclient==0.7.2
nbconvert==7.2.9
nbformat==5.7.3
nest-asyncio==1.5.6
netifaces==0.11.0
networkx==3.0
nn==0.1.1
notebook==6.5.2
notebook_shim==0.2.2
numpy==1.23.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.2
olefile==0.46
opencv-python==4.7.0.68
opt-einsum==3.3.0
outcome==1.3.0.post0
packaging==22.0
pandas==2.0.1
pandocfilters==1.5.0
paramiko==2.9.3
parse==1.20.2
parse_type==0.6.3
parso==0.8.3
pexpect==4.8.0
pickle5==0.0.11
pickleshare==0.7.5
Pillow==9.4.0
pillow-heif==0.9.3
platformdirs==2.5.1
pluggy==1.5.0
ply==3.11
prometheus-client==0.16.0
prompt-toolkit==3.0.36
protobuf==4.23.0
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio==0.2.14
PyAutoGUI==0.9.53
pycairo==1.20.1
PyChromecast==9.4.0
pycparser==2.21
pycups==2.0.1
pydicom==2.3.1
PyGetWindow==0.0.9
Pygments==2.14.0
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
pymongo==4.8.0
PyMsgBox==1.0.9
PyNaCl==1.5.0
pyparsing==2.4.7
pyperclip==1.8.2
PyRect==0.2.0
pyRFC3339==1.1
pyrsistent==0.19.3
PyScreeze==0.1.28
PySocks==1.7.1
pytest==8.3.2
pytest-bdd==7.2.0
pytest-html==4.1.1
pytest-metadata==3.1.1
python-apt==2.4.0+ubuntu4
python-dateutil==2.8.2
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.0.1
python-json-logger==2.0.4
python3-xlib==0.15
pythran==0.10.0
pyttsx3==2.90
pytweening==1.0.4
pytz==2022.1
PyWavelets==1.4.1
pyxdg==0.27
PyYAML==5.4.1
pyzmq==25.0.0
qtconsole==5.4.0
QtPy==2.3.0
regex==2022.10.31
reportlab==3.6.8
requests==2.28.1
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rsa==4.9
scikit-image==0.19.3
scikit-learn==1.2.2
scipy==1.8.0
scour==0.38.2
SecretStorage==3.3.1
selenium==4.24.0
Send2Trash==1.8.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
SpeechRecognition==3.10.4
sqlparse==0.2.4
ssh-import-id==5.11
stack-data==0.6.2
systemd-python==234
tensorboard==2.12.3
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorflow==2.12.0
tensorflow-estimator==2.11.0
tensorflow-io-gcs-filesystem==0.29.0
termcolor==2.2.0
terminado==0.17.1
texttable==1.6.4
tf-slim==1.1.0
tflearn==0.5.0
threadpoolctl==3.1.0
tifffile==2023.2.3
tinycss2==1.2.1
tomli==2.0.1
torch==1.13.1
tornado==6.2
tqdm==4.64.1
traitlets==5.8.1
trio==0.26.2
trio-websocket==0.11.1
ttystatus==0.38
typing_extensions==4.12.2
tzdata==2023.3
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
uri-template==1.2.0
urllib3==1.26.13
usb-creator==0.3.7
vboxapi==1.0
virtualenv==20.13.0+ds
wadllib==1.3.6
wcwidth==0.2.5
webcolors==1.12
webdriver-manager==4.0.2
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==2.2.2
widgetsnbextension==4.0.5
WoeUSB-ng==0.2.10.post2
wrapt==1.14.1
wsproto==1.2.0
wxPython==4.0.7
xdg==5
xkit==0.0.0
zeroconf==0.38.3
zipp==1.0.0
